Make sdk class/dependencies work in Android.
Part of the Java Client APIs responsibilities should include working
in Android.  HttpClient usage was removed since it conflicts with the
HttpClient that comes with the Android Standard API, and
commons-codec methods are only used if they exist in version 1.2,
which Android also includes in its Standard API.

Also included is a sample Android app for testing things.  Right now,
all that is included is a LoginActivity that, upon successful login,
switched to a basic Activity with a text field.