Create md5 string with commons-codec Hex.
This is the source of the SIDECAR-409 bug.
io.sidecar.security.SecurityUtils#md5 was generating MD5 strings that
were stripping leading 0s from the String.  This is due to our
conversion of the md5 byte[] first to a BigInteger and then envoking
toString(16) on the BigInteger instance.

The quickest route to fix this was to bring in another dependency on
commons-codec just so that we can quickly get this working on the
server side.